Re-connect the info page with next.js.
