--- web/app/info/info.css
@@ -1,5 +1,0 @@
-h1 {
-  background-image: linear-gradient(45deg, white, #67D864);
-  background-clip: text;
-  color: transparent;
-}
