+++ web/next.config.ts
@@ -3,6 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   output: "export", // Output static HTML/JS.
   distDir: "build",
+  trailingSlash: true,
   webpack: (config, options) => {
     config.module.rules.push({
       test: /\.vert$|\.frag$/,
